Add files via upload
In the spike trains folder you will ﬁnd the data ﬁle rho.dat. This is contains data collected and provided by Rob de Ruyter van Steveninck from a ﬂy H1 neuron responding to an approximate white-noise visual motion stimulus. Data were collected for 20 minutes at a sampling rate of 500 Hz. In the ﬁle, rho is a vector that gives the sequence of spiking events or nonevents at the sampled time, that is, every 2 ms. Simple programmes, load.jl and load.py, are supplied to load this data ﬁle and the data ﬁle described in the next question. Calculate the Fano factor and coeﬃcient of variation for this spike train as for the simulated spike trains above.